feat: v2 updates

+++ src/SoftwareSerialInputStream/SoftwareSerialInputStream.cpp
@@ -14,7 +14,7 @@ int SoftwareSerialInputStream::read() {
 }
 
 int SoftwareSerialInputStream::read(unsigned char *b, int off, int len) {
-	return softwareSerial->readBytes(&b[off], len);
+	return (int) softwareSerial->readBytes(&b[off], len);
 }
 
 #endif// OZEROIO_IO_SOFTWARE_SERIAL_SUPPORT_ENABLED
